Continue setup of wordpress website into seperate logical template files.
Add template tags to the following template files:
- header.php
- content.php and content-none.php

Create template files content.php and content-none.php. The first is used
to display content and contains conditionals within the file to determine whether
to dislpay the post title with anchor tags or not. Also a conditional is included in
content.php to determine whether to display an excerpt of a posts' content or the
full content. Template file `content-none.php` on the oter hand simply outputs an
error message using Twitter Bootstrap alert css component - may potentially be used
to display errors.