Add a sidebar/widget area
Does not include a conditional to check if any widgets are currently active which may
be a good feature to implement at a later time.

Include blog.css which goes with the Bootstrap Template used as a base for this theme,
cut down on time for making the site look good. In addition, a main style sheet that
came with this boilerplate was hooked into the functions.php so that it occurs after all
the bootstrap styles and as mentioned the blog.css. This main.css will mostly override styles
decalred before it.

Add template tags to the footer.php, this outputs minor details such as the WP version
being run for this WP install.