Add styles to main.css after adding a header navigation menu to the header template file. The styles try to acheive the same look and feel as the default blog.css used for the TWBS template of the same name. However when the default functionality of the wp_nav_menu template tag will wrap these items in a unordered list html tag which is different than the template which did not wrap in an unordered list and instead consisted of just anchor tags. May add this feature at a later time to have the same markup be outputted as the template had yet with using wordpress nav menu template tags.
Add functions to the functions.php to register the header navigation menu and
additionally an extra navigation menu which has yet to be placed in a template
file/partial to be displayed. Will likely display this extra nav menu at a
later date.

Remove the link and script HTML tags from the header.php and footer.php
template partials in order to dynamically enqueue them using the appropriate
functions.php file.